Linux support

+++ library/common/timer.h
@@ -6,11 +6,11 @@ struct xTimer_t;
 struct xTimer_t* Timer_Alloc(void);
 void Timer_Free(struct xTimer_t* pxTimer);
 
-float Timer_GetElapsedMicroSeconds(struct xTimer_t* pxTimer);
-float Timer_GetElapsedMilliSeconds(struct xTimer_t* pxTimer);
-float Timer_GetElapsedSeconds(struct xTimer_t* pxTimer);
-float Timer_GetDeltaTime(struct xTimer_t* pxTimer);
-float Timer_GetTime(struct xTimer_t* pxTimer);
+double Timer_GetElapsedMicroSeconds(struct xTimer_t* pxTimer);
+double Timer_GetElapsedMilliSeconds(struct xTimer_t* pxTimer);
+double Timer_GetElapsedSeconds(struct xTimer_t* pxTimer);
+double Timer_GetDeltaTime(struct xTimer_t* pxTimer);
+double Timer_GetTime(struct xTimer_t* pxTimer);
 
 void Timer_Start(struct xTimer_t* pxTimer);
 void Timer_Measure(struct xTimer_t* pxTimer);
